[Fix] crash with heredoc in case of invalid cmd
The variable heredoc_filename was garbage if the cmd was invalid.
The shutdown of the heredocs was causing the crash by freeing the
garbage

+++ src/parser/heredoc/heredoc_shutdown.c
@@ -22,6 +22,8 @@ void	parser_heredoc_shutdown(t_array *heredocs)
 	while (heredocs->len)
 	{
 		array_pop(heredocs, &redir);
+		if (redir->heredoc_filename == NULL)
+			continue ;
 		unlink(redir->heredoc_filename);
 		free(redir->heredoc_filename);
 	}
